requirements.txt
Pillow
torch
fastapi
uvicorn
pydantic
slowapi
environs
